services/notification-service/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
jinja2>=3.1.2
aiosmtplib>=3.0.1
email-validator>=2.1.0
pydantic>=2.6.0
httpx>=0.26.0
